Add tests related to noise simulator
