+++ test/unit_tests/braket/default_simulator/test_gate_operations.py
@@ -15,7 +15,7 @@ import pytest
 
 import braket.ir.jaqcd as instruction
 from braket.default_simulator import gate_operations
-from braket.default_simulator.operation_helpers import check_unitary
+from braket.default_simulator.operation_helpers import check_unitary, from_braket_instruction
 from braket.ir.jaqcd import shared_models
 
 testdata = [
@@ -76,16 +76,16 @@ testdata = [
 
 @pytest.mark.parametrize("instruction, targets, operation_type", testdata)
 def test_gate_operation_matrix_is_unitary(instruction, targets, operation_type):
-    check_unitary(gate_operations.from_braket_instruction(instruction).matrix)
+    check_unitary(from_braket_instruction(instruction).matrix)
 
 
 @pytest.mark.parametrize("instruction, targets, operation_type", testdata)
 def test_from_braket_instruction(instruction, targets, operation_type):
-    operation_instance = gate_operations.from_braket_instruction(instruction)
+    operation_instance = from_braket_instruction(instruction)
     assert isinstance(operation_instance, operation_type)
     assert operation_instance.targets == targets
 
 
 @pytest.mark.xfail(raises=ValueError)
 def test_from_braket_instruction_unsupported_instruction():
-    gate_operations.from_braket_instruction(shared_models.DoubleTarget(targets=[4, 3]))
+    from_braket_instruction(shared_models.DoubleTarget(targets=[4, 3]))
